Move libm checking to required library handling and invert logic.
We _need_ math functions, so libm is functionally a required library on
platforms that need it, thus it makes logical sense to handle it here even
though it’s auto-detected.

Additionaly, the common case for platforms we build on is that libm is
required, so it makes more sense to check if building with it works or
not and then fall back to checking for required functions that would
normally need libm, thus avoiding a pointless check in the common case.

Also, CMP0075 is implicitly set to NEW by cmake_minimum_required(3.13),
so there is no reason to set it ourselves.