Detect existing installs from running agent path.
This adds a fallback case in the existing install detection code in the
kickstart script that will attempt to find an existing install from a
running instance of the netdata agent. This works by checking for the
first process with a comm of `netdata` and determining the full path to
it’s 0th argument, then processing this just like we do for a detected
netdata agent path.

Additionally, this updates the part of the claiming code in the script
that handles determining if the agent is running to better take advantage
of the information from detecting an existing install, which should make
behavior a bit more consistent.